Revert "Renamed component from ProfileWallet to WalletOverview"
This reverts commit 542350233406f47aa5bf018f2dd2ae757025c357.
