+++ pages/wallet/index.vue
@@ -1,19 +1,19 @@
 <template>
   <page-template>
-    <wallet-overview class="col-md-10 mx-auto" />
+    <ProfileWallet class="col-md-10 mx-auto" />
     <wallet-tabs class="mt-8 col-md-10" />
   </page-template>
 </template>
 
 <script>
 import PageTemplate from "@/components/PageTemplate";
-import WalletOverview from "~/components/Wallet/WalletOverview.vue";
+import ProfileWallet from "~/components/Wallet/ProfileWallet.vue";
 import WalletTabs from "~/components/Wallet/WalletTabs/WalletTabs.vue";
 
 export default {
   components: {
     PageTemplate,
-    WalletOverview,
+    ProfileWallet,
     WalletTabs
   },
